File rename: live-android-service-setup-business-logic.ts -> android-service-configurator.ts
