+++ src/electron/platform/android/setup/android-service-configurator.ts
