+++ src/tests/unit/tests/electron/platform/android/setup/live-android-service-setup-business-logic-factory.test.ts
@@ -6,7 +6,7 @@ import { AdbWrapper, AdbWrapperFactory } from 'electron/platform/android/adb-wra
 import {
     AndroidServiceSetupBusinessLogic,
     LiveAndroidServiceSetupBusinessLogic,
-} from 'electron/platform/android/setup/live-android-service-setup-business-logic';
+} from 'electron/platform/android/setup/android-service-configurator';
 import { LiveAndroidServiceSetupBusinessLogicFactory } from 'electron/platform/android/setup/live-android-service-setup-business-logic-factory';
 import { IMock, Mock, MockBehavior, Times } from 'typemoq';
